refactor: move system event handlers to new `sys_event_handlers` module, and rename AdviceInjector
