+++ assembly/src/ast/tests.rs
@@ -467,14 +467,14 @@ fn test_ast_parsing_adv_ops() -> Result<(), Report> {
 
 #[test]
 fn test_ast_parsing_adv_injection() -> Result<(), Report> {
-    use super::AdviceInjectorNode::*;
+    use super::SystemEventNode::*;
 
     let context = TestContext::new();
     let source = source_file!(&context, "begin adv.push_u64div adv.push_mapval adv.insert_mem end");
     let forms = module!(begin!(
-        inst!(AdvInject(PushU64Div)),
-        inst!(AdvInject(PushMapVal)),
-        inst!(AdvInject(InsertMem))
+        inst!(SysEvent(PushU64Div)),
+        inst!(SysEvent(PushMapVal)),
+        inst!(SysEvent(InsertMem))
     ));
     assert_eq!(context.parse_forms(source)?, forms);
     Ok(())
