+++ core/src/sys_events.rs
@@ -1,12 +1,13 @@
 use core::fmt;
-use super::SignatureKind;
 
-// ADVICE INJECTORS
+// SYSTEM EVENTS
 // ================================================================================================
 
-// Randomly generated constant values for the VM's native events. All values were sampled
+// Randomly generated constant values for the VM's system events. All values were sampled
 // between 0 and 2^32.
 pub use constants::*;
+
+use super::SignatureKind;
 #[rustfmt::skip]
 mod constants {
     pub const EVENT_MERKLE_NODE_MERGE: u32            = 276124218;
@@ -38,8 +39,8 @@ mod constants {
 /// All actions, except for `MerkleNodeMerge`, `Ext2Inv` and `UpdateMerkleNode` can be invoked
 /// directly from Miden assembly via dedicated instructions.
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub enum AdviceInjector {
-    // MERKLE STORE INJECTORS
+pub enum SystemEvent {
+    // MERKLE STORE EVENTS
     // --------------------------------------------------------------------------------------------
     /// Creates a new Merkle tree in the advice provider by combining Merkle trees with the
     /// specified roots. The root of the new tree is defined as `Hash(LEFT_ROOT, RIGHT_ROOT)`.
@@ -56,7 +57,7 @@ pub enum AdviceInjector {
     /// provider (i.e., the input trees are not removed).
     MerkleNodeMerge,
 
-    // ADVICE STACK INJECTORS
+    // ADVICE STACK SYSTEM EVENTS
     // --------------------------------------------------------------------------------------------
     /// Pushes a node of the Merkle tree specified by the values on the top of the operand stack
     /// onto the advice stack.
@@ -226,7 +227,7 @@ pub enum AdviceInjector {
     ///   Advice stack: [ilog2(n), ...]
     ILog2,
 
-    // ADVICE MAP INJECTORS
+    // ADVICE MAP SYSTEM EVENTS
     // --------------------------------------------------------------------------------------------
     /// Reads words from memory at the specified range and inserts them into the advice map under
     /// the key `KEY` located at the top of the stack.
@@ -301,64 +302,64 @@ pub enum AdviceInjector {
     FalconSigToStack,
 }
 
-impl AdviceInjector {
+impl SystemEvent {
     pub fn into_event_id(self) -> u32 {
         match self {
-            AdviceInjector::MerkleNodeMerge => EVENT_MERKLE_NODE_MERGE,
-            AdviceInjector::MerkleNodeToStack => EVENT_MERKLE_NODE_TO_STACK,
-            AdviceInjector::MapValueToStack => EVENT_MAP_VALUE_TO_STACK,
-            AdviceInjector::MapValueToStackN => EVENT_MAP_VALUE_TO_STACK_N,
-            AdviceInjector::U64Div => EVENT_U64_DIV,
-            AdviceInjector::Ext2Inv => EVENT_EXT2_INV,
-            AdviceInjector::Ext2Intt => EVENT_EXT2_INTT,
-            AdviceInjector::SmtPeek => EVENT_SMT_PEEK,
-            AdviceInjector::U32Clz => EVENT_U32_CLZ,
-            AdviceInjector::U32Ctz => EVENT_U32_CTZ,
-            AdviceInjector::U32Clo => EVENT_U32_CLO,
-            AdviceInjector::U32Cto => EVENT_U32_CTO,
-            AdviceInjector::ILog2 => EVENT_ILOG2,
-            AdviceInjector::MemToMap => EVENT_MEM_TO_MAP,
-            AdviceInjector::HdwordToMap => EVENT_HDWORD_TO_MAP,
-            AdviceInjector::HdwordToMapWithDomain => EVENT_HDWORD_TO_MAP_WITH_DOMAIN,
-            AdviceInjector::HpermToMap => EVENT_HPERM_TO_MAP,
-            AdviceInjector::FalconSigToStack => EVENT_FALCON_SIG_TO_STACK,
+            SystemEvent::MerkleNodeMerge => EVENT_MERKLE_NODE_MERGE,
+            SystemEvent::MerkleNodeToStack => EVENT_MERKLE_NODE_TO_STACK,
+            SystemEvent::MapValueToStack => EVENT_MAP_VALUE_TO_STACK,
+            SystemEvent::MapValueToStackN => EVENT_MAP_VALUE_TO_STACK_N,
+            SystemEvent::U64Div => EVENT_U64_DIV,
+            SystemEvent::Ext2Inv => EVENT_EXT2_INV,
+            SystemEvent::Ext2Intt => EVENT_EXT2_INTT,
+            SystemEvent::SmtPeek => EVENT_SMT_PEEK,
+            SystemEvent::U32Clz => EVENT_U32_CLZ,
+            SystemEvent::U32Ctz => EVENT_U32_CTZ,
+            SystemEvent::U32Clo => EVENT_U32_CLO,
+            SystemEvent::U32Cto => EVENT_U32_CTO,
+            SystemEvent::ILog2 => EVENT_ILOG2,
+            SystemEvent::MemToMap => EVENT_MEM_TO_MAP,
+            SystemEvent::HdwordToMap => EVENT_HDWORD_TO_MAP,
+            SystemEvent::HdwordToMapWithDomain => EVENT_HDWORD_TO_MAP_WITH_DOMAIN,
+            SystemEvent::HpermToMap => EVENT_HPERM_TO_MAP,
+            SystemEvent::FalconSigToStack => EVENT_FALCON_SIG_TO_STACK,
         }
     }
 
-    /// Returns an advice injector corresponding to the specified event ID, or `None` if the event
+    /// Returns a system event corresponding to the specified event ID, or `None` if the event
     /// ID is not recognized.
     pub fn from_event_id(event_id: u32) -> Option<Self> {
         match event_id {
-            EVENT_MERKLE_NODE_MERGE => Some(AdviceInjector::MerkleNodeMerge),
-            EVENT_MERKLE_NODE_TO_STACK => Some(AdviceInjector::MerkleNodeToStack),
-            EVENT_MAP_VALUE_TO_STACK => Some(AdviceInjector::MapValueToStack),
-            EVENT_MAP_VALUE_TO_STACK_N => Some(AdviceInjector::MapValueToStackN),
-            EVENT_U64_DIV => Some(AdviceInjector::U64Div),
-            EVENT_EXT2_INV => Some(AdviceInjector::Ext2Inv),
-            EVENT_EXT2_INTT => Some(AdviceInjector::Ext2Intt),
-            EVENT_SMT_PEEK => Some(AdviceInjector::SmtPeek),
-            EVENT_U32_CLZ => Some(AdviceInjector::U32Clz),
-            EVENT_U32_CTZ => Some(AdviceInjector::U32Ctz),
-            EVENT_U32_CLO => Some(AdviceInjector::U32Clo),
-            EVENT_U32_CTO => Some(AdviceInjector::U32Cto),
-            EVENT_ILOG2 => Some(AdviceInjector::ILog2),
-            EVENT_MEM_TO_MAP => Some(AdviceInjector::MemToMap),
-            EVENT_HDWORD_TO_MAP => Some(AdviceInjector::HdwordToMap),
-            EVENT_HDWORD_TO_MAP_WITH_DOMAIN => Some(AdviceInjector::HdwordToMapWithDomain),
-            EVENT_HPERM_TO_MAP => Some(AdviceInjector::HpermToMap),
-            EVENT_FALCON_SIG_TO_STACK => Some(AdviceInjector::FalconSigToStack),
+            EVENT_MERKLE_NODE_MERGE => Some(SystemEvent::MerkleNodeMerge),
+            EVENT_MERKLE_NODE_TO_STACK => Some(SystemEvent::MerkleNodeToStack),
+            EVENT_MAP_VALUE_TO_STACK => Some(SystemEvent::MapValueToStack),
+            EVENT_MAP_VALUE_TO_STACK_N => Some(SystemEvent::MapValueToStackN),
+            EVENT_U64_DIV => Some(SystemEvent::U64Div),
+            EVENT_EXT2_INV => Some(SystemEvent::Ext2Inv),
+            EVENT_EXT2_INTT => Some(SystemEvent::Ext2Intt),
+            EVENT_SMT_PEEK => Some(SystemEvent::SmtPeek),
+            EVENT_U32_CLZ => Some(SystemEvent::U32Clz),
+            EVENT_U32_CTZ => Some(SystemEvent::U32Ctz),
+            EVENT_U32_CLO => Some(SystemEvent::U32Clo),
+            EVENT_U32_CTO => Some(SystemEvent::U32Cto),
+            EVENT_ILOG2 => Some(SystemEvent::ILog2),
+            EVENT_MEM_TO_MAP => Some(SystemEvent::MemToMap),
+            EVENT_HDWORD_TO_MAP => Some(SystemEvent::HdwordToMap),
+            EVENT_HDWORD_TO_MAP_WITH_DOMAIN => Some(SystemEvent::HdwordToMapWithDomain),
+            EVENT_HPERM_TO_MAP => Some(SystemEvent::HpermToMap),
+            EVENT_FALCON_SIG_TO_STACK => Some(SystemEvent::FalconSigToStack),
             _ => None,
         }
     }
 }
 
-impl crate::prettier::PrettyPrint for AdviceInjector {
+impl crate::prettier::PrettyPrint for SystemEvent {
     fn render(&self) -> crate::prettier::Document {
         crate::prettier::display(self)
     }
 }
 
-impl fmt::Display for AdviceInjector {
+impl fmt::Display for SystemEvent {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             Self::MerkleNodeMerge => write!(f, "merkle_node_merge"),
